+++ lib/settings/widgets/section_session.dart
@@ -1,19 +1,56 @@
+/// Session section.
+//
+// Time-stamp: <Monday 2025-01-06 15:20:25 +1100 Graham Williams>
+//
+/// Copyright (C) 2024, Togaware Pty Ltd
+///
+/// Licensed under the GNU General Public License, Version 3 (the "License");
+///
+/// License: https://www.gnu.org/licenses/gpl-3.0.en.html
+//
+// This program is free software: you can redistribute it and/or modify it under
+// the terms of the GNU General Public License as published by the Free Software
+// Foundation, either version 3 of the License, or (at your option) any later
+// version.
+//
+// This program is distributed in the hope that it will be useful, but WITHOUT
+// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+// FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+// details.
+//
+// You should have received a copy of the GNU General Public License along with
+// this program.  If not, see <https://www.gnu.org/licenses/>.
+///
+/// Authors: Kevin Wang
+import 'dart:io';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:markdown_tooltip/markdown_tooltip.dart';
 import 'package:rattle/constants/spacing.dart';
 import 'package:rattle/providers/session_control.dart';
 import 'package:rattle/providers/settings.dart';
-import 'package:rattle/settings/utils/session_settings.dart';
+import 'package:rattle/settings/utils/save_image_viewer_app.dart';
 import 'package:rattle/settings/widgets/image_viewer_text_field.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class Session extends ConsumerWidget {
   const Session({super.key});
 
-  Future<void> _saveAskOnExit(bool value) async {
+  /// Save ask on exit setting
+  Future<void> saveAskOnExit(bool value) async {
     final prefs = await SharedPreferences.getInstance();
     await prefs.setBool('askOnExit', value);
+  }
+
+  /// Reset session control settings
+  void resetSessionControl(WidgetRef ref) {
+    ref.invalidate(askOnExitProvider);
+    saveAskOnExit(true);
+
+    final defaultApp = Platform.isWindows ? 'start' : 'open';
+    ref.read(imageViewerSettingProvider.notifier).state = defaultApp;
+    saveImageViewerApp(defaultApp);
   }
 
   @override
@@ -80,7 +117,7 @@ class Session extends ConsumerWidget {
                     value: askOnExit,
                     onChanged: (value) {
                       ref.read(askOnExitProvider.notifier).state = value;
-                      _saveAskOnExit(value);
+                      saveAskOnExit(value);
                     },
                   ),
                 ],
